Fixes for C# Decimal conversions
git-svn-id: svn://gil-w2k8/HexEdit_Repository/trunk@983 295070fd-a27b-da4f-a32e-77c6a882a0e3

+++ HexEdit/BigInteger.cpp
@@ -437,7 +437,7 @@ int BigInteger::Compare(const BigInteger & Lhs, const BigInteger & Rhs)
 
 	// If SigBit = -1, i = 0
 	int i = (Lhs.SigBit < 0) ? (- Lhs.SigBit - 2) : (Lhs.SigBit - 1);
-	i /= (sizeof(Unit) * 8);
+	i /= (int)(sizeof(Unit) * 8);
 
 	for (; i >= 0; --i)
 	{
